app/requirements.txt
Django>=4.2.3
gunicorn>=21.2.0
psycopg2-binary>=2.9.6
djangorestframework>=3.12.2
django-filter>=2.4.0
django-environ>=0.4.5
django-cors-headers>=3.7.0
PyJWT>=2.10.1
django-webpack-loader>=0.6.0
whitenoise>=5.2.0
dj-database-url>=0.5.0
requests>=2.31.0